pyproject.toml
[tool.poetry]
name = "pokemon"
version = "0.1.0"
description = ""
authors = ["Thomas Legrand <legrand.thomas5@hotmail.fr>"]

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^0.25.3"
streamlit = "^0.51.0"
plotly_express = "^0.4.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.4"
pylint = "^2.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
